Use makefile instead of cake
I like cake as it has lots of builtin libraries.
But that's just another thing to learn. make is quite synonymous for build tool.

Though I'm not really use to it but there seems to be quite a fair bit of help in the wild.

Till i find another agnostic popular build tool we will migrate to makefile