use terraform cloud to apply to infrastructure
Unfortunately terraform cloud cannot output files. You'll need to use terraform show
but that requires to write over an existing file which i can't do.

So perhaps using just whatever is in the cloud should work?
Not sure how this would work if multiple PR happens for the terraform files