redesign(dotnet): Move the logic back into the domain
This causes workflow to be async but I think that's a small worry compared
to the logic of where the behaviour to detect an existing recipe.
DataAccess shouldn't have business logic per-se; it should only understand
how to retrieve data and save data.

Knowing if the recipe existed or not, is a domain logic concern. So rightfully
it has gone back into the Domain